+++ solver/src/main/java/org/chocosolver/solver/search/restart/LinearCutoff.java
@@ -28,7 +28,7 @@ public class LinearCutoff extends AbstractCutoff {
     @Override
     public long getNextCutoff() {
         calls++;
-        return scaleFactor * calls;
+        return scaleFactor * calls * grower.getAsInt();
     }
 
     @Override