+++ solver/src/test/java/org/chocosolver/solver/search/loop/SolverTest.java
@@ -9,7 +9,6 @@
  */
 package org.chocosolver.solver.search.loop;
 
-import org.chocosolver.solver.search.restart.LubyCutoff;
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.Settings;
 import org.chocosolver.solver.Solver;
@@ -17,8 +16,10 @@ import org.chocosolver.solver.constraints.Constraint;
 import org.chocosolver.solver.propagation.PropagationProfiler;
 import org.chocosolver.solver.search.limits.NodeCounter;
 import org.chocosolver.solver.search.loop.lns.neighbors.RandomNeighborhood;
+import org.chocosolver.solver.search.loop.monitors.SolvingStatisticsFlow;
 import org.chocosolver.solver.search.loop.move.MoveBinaryDFS;
 import org.chocosolver.solver.search.loop.move.MoveBinaryLDS;
+import org.chocosolver.solver.search.restart.LubyCutoff;
 import org.chocosolver.solver.search.strategy.Search;
 import org.chocosolver.solver.search.strategy.decision.Decision;
 import org.chocosolver.solver.search.strategy.strategy.AbstractStrategy;
@@ -26,12 +27,15 @@ import org.chocosolver.solver.search.strategy.strategy.FullyRandom;
 import org.chocosolver.solver.search.strategy.strategy.IntStrategy;
 import org.chocosolver.solver.variables.BoolVar;
 import org.chocosolver.solver.variables.IntVar;
+import org.chocosolver.util.ProblemMaker;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.io.PrintWriter;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import static org.chocosolver.solver.search.strategy.Search.*;
 import static org.chocosolver.util.ProblemMaker.makeGolombRuler;
@@ -402,5 +406,31 @@ public class SolverTest {
                 " 7              1         0         0         0         0  \"binLoad[1]\"\n" +
                 " 8              1         0         0         0         0  \"binLoad[2]\"\n" +
                 " 9              0         0         0         0         0  \"cste -- 24576\"\n\n");
+    }
+
+    @Test(groups = "1s")
+    public void testSolvingFlow() {
+        Model model = ProblemMaker.makeGolombRuler(9);
+        Solver solver = model.getSolver();
+        String jsonObject = SolvingStatisticsFlow.toJSON(solver);
+        Pattern p = Pattern.compile("\"variables\":\"(\\d+)\"," +
+                "\"constraints\":\"(\\d+)\"," +
+                "\"objective\":\"(\\d+)\"," +
+                "\"solutions\":\"(\\d+)\"," +
+                "\"nodes\":\"(\\d+)\"," +
+                "\"fails\":\"(\\d+)\"," +
+                "\"backtracks\":\"(\\d+)\"," +
+                "\"backjumps\":\"(\\d+)\"," +
+                "\"restarts\":\"(\\d+)\"," +
+                "\"fixpoints\":\"(\\d+)\"," +
+                "\"depth\":\"(\\d+)\"," +
+                "\"time\":\"(\\d+):(\\d+):(\\d+)\"," +
+                "\"memory\":\"-?(\\d+)\"}"
+        );
+        Matcher m = p.matcher(jsonObject);
+        Assert.assertTrue(m.find());
+        Assert.assertEquals(m.group(1), "45");
+        Assert.assertEquals(m.group(2), "155");
+        Assert.assertEquals(m.group(3), "1024");
     }
 }